Add files via upload
Multiple challenges in to defining numbers in a given range, in order to find manufacturing faults in a fictional company.

door_numbers.py - A company tasked with producing door numbers for a set of new houses has had a manufacturing fault with the number 7's, please find out how many number 7's they will need to make in a range of 1-250 homes.

door_numbers_2.py - The same company from the first challenge has been tasked with creating door numbers for a hotel which has 1150 rooms. A manufacturing fault has lead to all doors containing the numbers 6 needing replaced. The numbers also double up in usage for the number 9, please check how many units will need to be manufactured for the hotel.